feat: mingletek
因應需求做調整：
 正是環境調整
修改程式

+++ common/model/SubPicture.php
@@ -34,7 +34,7 @@ class SubPicture extends Model
 	 * @param int $confidence
 	 * @return bool
 	 */
-	public static function modifyByStoreIdAndPicture(int $storeId, string $picture, int $productId, int $confidence)
+	public static function modifyByStoreIdAndPicture(int $storeId, string $picture, int $productId, float $confidence)
 	{
 		return self::getDb()->update(
 			[
